Backend/requirements.txt
bidict==0.21.2
dnspython==1.16.0
eventlet==0.30.2
greenlet==1.1.0
gunicorn==20.1.0
pkg-resources==0.0.0
python-engineio==4.2.0
python-socketio==5.3.0
six==1.16.0
